API Reference

Original DRSOM with 2 directions $g_k$ and $d_k$

DRSOM.DRSOMIteration — Type

Iteration object for DRSOM, to initialize an iterator, you must specify the attributes, including

attr	notes

x0	initial point

f	the smooth function to minimize

ϕ	the nonsmooth function

g	the gradient function

ga	gradient function via forward or backward diff

hvp	hvp function via forward or backward diff

H	hessian function

rest of the attributes have default options:

    t::Dates.DateTime = Dates.now()
    itermax::Int64 = 20
    fog = :forward
    sog = :forward

Note

We do not currently support nonsmooth optimization, this is a future plan.

DRSOM.DRSOMState — Type

State struct for DRSOM to keep the iterate information, We keep the following attributes:

attr	notes

x::Tx	iterate

fx::R	new value f at x: x(k)

fz::R	old value f at z: x(k-1)

∇f::Tx	gradient of f at x

∇fz::Tx	gradient of f at z

∇hvp::Tx	gradient buffer (for buffer use of hvps)

y::Tx	forward point

z::Tx	previous point

d::Tx	momentum/fixed-point diff at iterate (= x - z)

α₁::R	stepsize 1 parameter of gradient

α₂::R	stepsize 2 parameter of momentum

Q::Tq	Q for low-dimensional QP

c::Tc	c for low-dimensional QP

G::Tq	c for low-dimensional QP

Δ::R	trs radius

dq::R	decrease of estimated quadratic model

df::R	decrease of the real function value

ρ::R	trs descrease ratio: ρ = df/dq

ϵ::R	eps: residual for gradient

γ::R	scaling parameter γ for λ

ψ::R	1-D linear search iteration

λ::R	dual λ

kₜ::Int	inner iterations for adjustment

kf::Int	function evaluations

kg::Int	gradient evaluations

kh::Int	hvp evaluations

kH::Int	hessian evaluations

t::R	running time